Atan2, int return de polar (falta arreglar test)

+++ src/main/java/superior/ncom/ComplejoPolar.java
@@ -9,7 +9,7 @@ public class ComplejoPolar extends Complejo implements Mostrable{
         double sumaCuadrados = Math.pow(parteReal,2) + Math.pow(parteImaginaria,2);
         double r = Math.sqrt(sumaCuadrados);
         double phi = Math.atan(parteImaginaria / parteReal);
-        return "[ " + (int) r + " ; " + phi + "π ]";
+        return "[ " + r + " ; " + phi + "π ]";
     }
 
     public ComplejoBinomico transformarABinomico() {
